MINIFICPP-12: Resolve warnings found during compilation
MINIFICPP-12: Remove linux warnings

MINIFICPP-12: Update property and stringutils

+++ libminifi/include/utils/StringUtils.h
@@ -95,8 +95,8 @@ class StringUtils {
 
   static std::vector<std::string> split(const std::string &str, const std::string &delimiter) {
     std::vector<std::string> result;
-    int last = 0;
-    int next = 0;
+    size_t last = 0;
+    size_t next = 0;
     while ((next = str.find(delimiter, last)) != std::string::npos) {
       result.push_back(str.substr(last, next - last));
       last = next + delimiter.length();
